requirements.txt
easydict
thop
pyfasttext
faiss-cpu
mmengine
colour
pytorch_lightning==2.0
omegaconf==2.1.1
einops==0.3.0
editdistance==0.6.2
lmdb==1.4.0
matplotlib==3.6.2
opencv-python==4.9.0.80
Requests==2.31.0
ipython==8.7.0
pygame==2.5.2
pytest==7.4.3
test-tube==0.7.5
Pillow==9.5.0
scipy==1.11.4
imgaug==0.4.0
tensorboardX==2.6.2.2
setuptools==59.5.0
ptflops==0.7
basicsr==1.4.2
nltk==3.8.1
timm==0.9.16